+++ mlir/lib/Conversion/ToQIR/ToQIRPrep.cpp
@@ -1,8 +1,8 @@
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Dialect/LLVMIR/LLVMTypes.h"
 #include "mlir/IR/BuiltinOps.h"
+#include "qcc/Conversion/ToQIR/Constants.h"
 #include "qcc/Conversion/ToQIR/ToQIR.h"
-#include "qcc/Conversion/ToQIR/constants.h"
 
 using namespace mlir;
